Added Comman interfaces types
