+++ src/Types/Organization.ts
@@ -1,0 +1,19 @@
+export interface Organization {
+  _id: string;
+  org_name: string;
+  org_location: OrgLocation[];
+  org_created_at: string;
+  org_updated_at: string;
+  isActive: boolean;
+  __v: number;
+}
+
+// models/OrgLocation.ts
+export interface OrgLocation {
+  loc: string;
+  address: string;
+  loc_contact: number;
+  loc_email: string;
+  admin_id: string;
+  _id: string;
+}
