+++ src/Types/index.ts
@@ -1,0 +1,6 @@
+// models/index.ts
+
+export type { Organization } from "./Organization";
+export type { OrgLocation } from "./Organization";
+export type { RoleSpecificDetails } from "./User";
+export type { UserData  } from "./User";
